+++ click_extended/decorators/transform/__init__.py
@@ -29,7 +29,10 @@ from click_extended.decorators.transform.to_case import (
 )
 from click_extended.decorators.transform.to_date import to_date
 from click_extended.decorators.transform.to_datetime import to_datetime
+from click_extended.decorators.transform.to_directory import to_directory
+from click_extended.decorators.transform.to_file import to_file
 from click_extended.decorators.transform.to_path import to_path
+from click_extended.decorators.transform.to_symlink import to_symlink
 from click_extended.decorators.transform.to_time import to_time
 from click_extended.decorators.transform.to_timestamp import to_timestamp
 from click_extended.decorators.transform.truncate import truncate
@@ -64,7 +67,10 @@ __all__ = [
     "to_upper_case",
     "to_date",
     "to_datetime",
+    "to_directory",
+    "to_file",
     "to_path",
+    "to_symlink",
     "to_time",
     "to_timestamp",
     "truncate",